Refactor pyproject.toml and CI configuration; update import paths and dependencies
