+++ services/auth/pyproject.toml
@@ -51,3 +51,7 @@ exclude = "(alembic|.venv)"
 dev = [
     "pytest-cov>=6.2.1",
 ]
+
+[tool.pytest.ini_options]
+pythonpath = ["."]
+testpaths = ["tests"]
