+++ services/convertor/s3_client.py
@@ -5,7 +5,8 @@ from typing import AsyncGenerator, BinaryIO, Dict
 
 from aiobotocore.session import AioBaseClient, get_session
 from botocore.exceptions import ClientError
-from config import get_s3_settings
+
+from services.convertor.config import get_s3_settings
 
 settings = get_s3_settings()
 PART_SIZE = 1024 * 1024 * 10
