+++ services/convertor/tests/test_config.py
@@ -1,4 +1,4 @@
-from convertor.config import S3Settings, get_s3_settings
+from ..config import S3Settings, get_s3_settings
 
 
 def test_import_config() -> None:
